pyproject.toml
[project]
name = "personal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cmake>=3.30.5",
    "google-generativeai>=0.8.3",
    "loguru>=0.7.2",
    "mirascope[gemini,litellm,tenacity]>=1.6.0",
    "numpy>=2.1.2",
    "pyperclip>=1.9.0",
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "rerankers[api,flashrank]>=0.5.3",
    "rich>=13.9.3",
    "sentence-transformers>=3.2.1",
    "sqlite-vec>=0.1.3",
]
